Flattened platform command
