+++ src/commands.rs
@@ -1,3 +1,5 @@
+use lazy_static::lazy_static;
+use std::collections::HashMap;
 use std::env::{current_dir, set_current_dir};
 use std::fs::{create_dir_all, read_to_string, remove_file, write};
 use std::path::{Path, PathBuf};
@@ -11,9 +13,11 @@ use walkdir::WalkDir;
 
 use crate::build_utils::build_for_platform;
 use crate::config_utils::{
-    add_module_to_config, create_initial_config, get_config_as_object, is_module_in_config,
-    remove_module_from_config_if_exists,
+    add_module_to_config, add_platform_to_config, create_initial_config, get_config_as_object,
+    is_module_in_config, remove_module_from_config_if_exists,
+    remove_platform_from_config_if_exists,
 };
+use crate::cross_utils::add_image_override_for_platform;
 use crate::definitions::CargoToml;
 use crate::file_utils::write_and_fmt;
 use crate::gdnlib_utils::create_initial_gdnlib;
@@ -23,6 +27,18 @@ use crate::log_utils::{
     log_success_to_console, log_version, ConsoleColors,
 };
 use crate::path_utils::{exit_if_not_lib_dir, get_absolute_path};
+
+lazy_static! {
+    static ref VALID_PLATFORMS: HashMap<&'static str, &'static str> = {
+        let mut m = HashMap::new();
+        // m.insert("android.arm", "aarch64-linux-android");
+        // m.insert("android", "x86_64-linux-android");
+        m.insert("windows", "x86_64-pc-windows-gnu");
+        // m.insert("linux", "x86_64-unknown-linux-gnu");
+        // m.insert("macos", "x86_64-apple-darwin");
+        m
+    };
+}
 
 /// Creates the library used to manage Rust modules.
 ///
@@ -503,4 +519,93 @@ pub fn build_library_with_timestamp(is_release: bool, build_all_targets: bool) {
         "[{}] {}",
         current_datetime_formatted, "waiting for changes..."
     ));
+}
+
+/// Adds a new platform to the platforms that godot-rust-cli will build the
+/// library for.
+///
+/// Note that at this time godot-rust-cli only supports platforms with
+/// 64-bit architectures. If demand for 32-bit is desired it can be added but
+/// as of now only 64-bit is supported for any platform.
+///
+/// Platforms only need to be added if you are buliding for a different
+/// platform than your native platform. For example, if you are developing
+/// on Windows then you don't need to add Windows as a platfrom because the
+/// library will automatically be built for your native platform.
+///
+/// Any platform other than your native platform will be built using the
+/// `cross` crate from https://github.com/rust-embedded/cross. This means that
+/// in order to cross-compile, you will need to follow the instructions for
+/// setting it up which is essentially just installing the crate and making
+/// sure that you have docker or podman.
+///
+/// The list of platforms that can be provided are:
+/// Android
+/// Linux
+/// Windows
+/// MacOS
+///
+/// If you would like another platform to be added then please open an issue in
+/// the GitHub or let me know in the Discord.
+///
+/// # Arguments
+///
+/// `platform` - The platform to compile for.
+pub fn command_platform_add(platform: &str) {
+    let platform_normalized = platform.to_lowercase();
+
+    if VALID_PLATFORMS.contains_key(&platform_normalized.as_str()) {
+        let mut config = get_config_as_object();
+        // Add the platform to the `platforms` array in the config.
+        add_platform_to_config(&platform_normalized, &mut config);
+
+        // Since we need images that do more than the default cross images, we
+        // have to copy the docker file override into the user's library and
+        // add it
+        add_image_override_for_platform(&platform_normalized);
+    } else {
+        log_error_to_console(&format!("The target {} isn't a valid target. Please file an issue in the GitHub or Discord if this is incorrect.", &platform));
+        exit(1);
+    }
+}
+
+/// Removes a platform from the configuration.
+///
+/// # Arguments
+///
+/// `platform` - The platform to remove.
+pub fn command_platform_remove(platform: &str) {
+    let mut config = get_config_as_object();
+
+    // Remove the platform from the `platforms` array in the configuration.
+    remove_platform_from_config_if_exists(platform, &mut config);
+
+    // Remove the docker image from the user's system since it's no longer
+    // needed.
+    let image_name = match platform {
+        "windows" => Some("godot-rust-cli-platform-windows:v1"),
+        _ => None,
+    };
+
+    match image_name {
+        Some(image_tag) => {
+            let mut remove_default_docker_image_command = std::process::Command::new("docker");
+            remove_default_docker_image_command
+                .arg("rmi")
+                .arg("rustembedded/cross:x86_64-pc-windows-gnu");
+
+            remove_default_docker_image_command
+                .status()
+                .expect("Unable to remove docker image rustembedded/cross:x86_64-pc-windows-gnu");
+
+            let mut remove_custom_docker_image_command = std::process::Command::new("docker");
+            remove_custom_docker_image_command.arg("rmi").arg(image_tag);
+
+            remove_custom_docker_image_command
+                .status()
+                .expect(&format!("Unable to remove docker image {}", image_tag));
+            log_info_to_console(&format!("Removed docker image for {}", &platform));
+        }
+        None => (),
+    }
 }
